requirements.txt
asgiref==3.5.2
Django==4.1
django-environ==0.9.0
mysqlclient==2.1.1
sqlparse==0.4.2
django-allauth==0.51.0
requests~=2.28.1